Implemented - Ball Escape Simulation
This commit implements all the functionality for the Ball Escape simulation I saw on tiktok and wanted to try out myself.

All the functionality was relativly simple to figure out, except for the ball bounces which took the most time (1 hour or so).

The music is still a work in progress as getting a song split up into proper notes is much harder than you would think.

Next, get a few notes to work with and implement any polish I need for this simulation before moving onto another.