Implemented - Basic MoreSpawnBigger Simulation
This commit starts work on the second simulation, it gets the basic implementations in with the escaped ball being removed and replaced with 3 new ones at the middle and the those new ones being bigger than before.

Next, fix the issues where the new balls don't bounce off of each other causing them to stay stationary at the bottom.