Implemented/Fixed - AdjustAllAudio Method + Simulation Issues
This commit implements the AdjustAllAudio method which turns off all the audio sources relevant to the simulation on click off the Audio Button (re-enabled upon 2nd click).

It also fixes the issues with the subsequent simulations not spawning any balls after the 1st escapes.

Next, implement a fix to set the audio on/off depending on the simulation audio button state.