Finished - 3rd Simulation Implementation
This commit finishes off the WhoWillWin simulation implementation. The third and final one for this project.

Next, that's it for what I wanted to do on this project, however, I may come back at a later date to improve upon the ball physics and possibly add an extra simulation or two.